+++ backend-rust/CHANGELOG.md
@@ -10,6 +10,7 @@ Database schema version: 18
 
 ### Added
 
+- Add support for ChainUpdatePayload events
 - Add query `SuspendedValidators::suspendedValidators`.
 - Add query `SuspendedValidators::primedForSuspensionValidators`.
 - Add database migration 18 adding table tracking baker pool and passive stake for every reward period.