+++ backend-rust/CHANGELOG.md
@@ -16,6 +16,7 @@ Database schema version: 20
 - Change `Versions::backend_versions` to `Versions::backend_version`.
 - When `effective_time` is zero then it translates into `Block::slot_time`.
 - Split up migration 18 into several SQL transactions to avoid timeouts for long running migrations.
+- Change CLI option `--log-level` to only apply for logs produced directly from this project, instead of including every dependency.
 
 ## [0.1.38] - 2025-03-21