+++ backend-rust/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-Database schema version: 20
+Database schema version: 22
 
 ### Added
 
@@ -13,6 +13,7 @@ Database schema version: 20
 
 ### Fixed
 
+- Add database migration 22 fixing some passive delegators that had no `delegated_restake_earnings` value set in the database.
 - Change `Versions::backend_versions` to `Versions::backend_version`.
 - When `effective_time` is zero then it translates into `Block::slot_time`.
 - Split up migration 18 into several SQL transactions to avoid timeouts for long running migrations.